gerenciador-demandas/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==2.0.2
aiosqlite==0.20.0
alabaster==0.7.13
altair==4.2.2
annotated-types==0.7.0
anyio==3.6.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
blinker==1.6.2
Brotli==1.1.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.2.1
cryptography==41.0.2
cssselect2==0.7.0
databases==0.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.8.0
dnspython==2.3.0
docutils==0.20.1
docx2pdf==0.1.8
ecdsa==0.18.0
email-validator==2.0.0.post2
entrypoints==0.4
env==0.1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-mail==1.4.1
fief-client==0.17.0
Flask==3.0.2
fonttools==4.42.1
fpdf2==2.7.5
frozenlist==1.4.1
gitdb==4.0.10
GitPython==3.1.31
gobject==0.1.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
imagesize==1.4.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==1.5.0
lxml==4.9.3
makefun==1.15.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
MouseInfo==0.1.3
multidict==6.0.5
numpy==1.24.3
openpyxl==3.1.2
orjson==3.8.11
packaging==23.1
pandas==2.0.1
passlib==1.7.4
pdfkit==1.0.0
pdfminer.six==20221105
Pillow==9.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.7.3
pydantic-settings==2.3.1
pydantic_core==2.18.4
pydyf==0.8.0
PyGetWindow==0.0.9
Pygments==2.15.1
PyJWT==2.8.0
PyMsgBox==1.0.9
PyPDF2==3.0.1
pyperclip==1.8.2
pyphen==0.14.0
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.30
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
pyttsx3==2.90
pytweening==1.0.7
pytz==2023.3
pywhatkit==5.4
PyYAML==6.0
reportlab==4.0.4
requests==2.31.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.1.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.49
starlette==0.37.2
tabula==1.0.5
tabula-py==2.9.0
tinycss2==1.2.1
toolz==0.12.0
tqdm==4.65.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.4
uvicorn==0.22.0
watchfiles==0.19.0
weasyprint==60.2
webencodings==0.5.1
websockets==11.0.2
Werkzeug==3.0.1
wikipedia==1.4.0
wrapt==1.15.0
yarl==1.9.4
zipp==3.15.0
zopfli==0.2.3
pytest==8.3.2
APScheduler==3.11.0
asyncpg==0.29.0